+++ examples/doa_algorithms.py
@@ -56,10 +56,11 @@ from pyroomacoustics.doa import circ_dist
 azimuth = 61.0 / 180.0 * np.pi  # 60 degrees
 distance = 3.0  # 3 meters
 dim = 2  # dimensions (2 or 3)
+room_dim = np.r_[10.0, 10.0]
 
 # Use AnechoicRoom or ShoeBox implementation. The results are equivalent because max_order=0 for both.
 # The plots change a little because in one case there are no walls.
-anechoic = True   
+use_anechoic_class = True   
 
 print("============ Using anechoic: {} ==================".format(anechoic))
 
@@ -74,18 +75,19 @@ freq_bins = np.arange(5, 60)  # FFT bins to use for estimation
 # compute the noise variance
 sigma2 = 10 ** (-SNR / 10) / (4.0 * np.pi * distance) ** 2
 
-# Create the experiment setup.
-# We use a circular array with radius 15 cm # and 12 microphones
-center = np.r_[5.0, 5.0]
-R = pra.circular_2D_array(center, 12, 0.0, 0.15)
-source_location = center + distance * np.r_[np.cos(azimuth), np.sin(azimuth)]
-source_signal = np.random.randn((nfft // 2 + 1) * nfft)
-
-if anechoic:
+# Create an anechoic room
+if use_anechoic_class:
     aroom = pra.AnechoicRoom(dim, fs=fs, sigma2_awgn=sigma2)
 else:
-    aroom = pra.ShoeBox(2 * center, fs=fs, max_order=0, sigma2_awgn=sigma2)
+    aroom = pra.ShoeBox(room_dim, fs=fs, max_order=0, sigma2_awgn=sigma2)
+
+# add the source
+source_location = room_dim / 2 + distance * np.r_[np.cos(azimuth), np.sin(azimuth)]
+source_signal = np.random.randn((nfft // 2 + 1) * nfft)
 aroom.add_source(source_location, signal=source_signal)
+
+# We use a circular array with radius 15 cm # and 12 microphones
+R = pra.circular_2D_array(room_dim / 2, 12, 0.0, 0.15)
 aroom.add_microphone_array(pra.MicrophoneArray(R, fs=aroom.fs))
 
 # run the simulation
@@ -118,8 +120,6 @@ for algo_name in algo_names:
     # doa.azimuth_recon contains the reconstructed location of the source
     print(algo_name)
     print("  Recovered azimuth:", doa.azimuth_recon / np.pi * 180.0, "degrees")
-    print(
-        "  Error:", circ_dist(azimuth, doa.azimuth_recon) / np.pi * 180.0, "degrees"
-    )
+    print("  Error:", circ_dist(azimuth, doa.azimuth_recon) / np.pi * 180.0, "degrees")
 
 plt.show()